+++ e2e-tests/browser/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @evervault/browser-e2e-tests
 
+## 1.0.37
+
+### Patch Changes
+
+- Updated dependencies [fc87e61]
+  - @evervault/browser@2.63.1
+
 ## 1.0.36
 
 ### Patch Changes