+++ e2e-tests/ui-components/vanilla-test-server/CHANGELOG.md
@@ -1,5 +1,12 @@
 # e2e-tests-ui-components-vanilla-server
 
+## 0.0.71
+
+### Patch Changes
+
+- Updated dependencies [fc87e61]
+  - @evervault/browser@2.63.1
+
 ## 0.0.70
 
 ### Patch Changes